+++ CHANGELOG.md
@@ -10,9 +10,12 @@
 
 - Renamed `<Sparkbar />` to `<SparkBarChart />`.
 - Renamed `<Sparkline />` to `<SparkLineChart />`.
-- Removed `area`, `lineStyle`, and `hasPoint` from `SparkLineChartProps.series`.
 - Moved `offsetLeft` & `offsetRight` from `SparkLineChartProps.series` to `SparkLineChartProps`.
 - Renamed `series` to `data` in `SparkLineChartProps`.
+### Removed
+
+- `area`, `lineStyle`, and `hasPoint` from `SparkLineChartProps.series`.
+- `barColor` and `comparison` from `SparkBarChartProps`. These props are now available on the `DataSeries` as `color` and `isComparison`.
 
 ## [0.25.0] - 2021-11-12